Added Connection Feeds

+++ src/main/java/com/sqrfactor/service/FeedService.java
@@ -15,7 +15,7 @@ public interface FeedService {
 	
 	Feed findById(long feedId);
 	
-	List<Feed> findByUserId(int userId);
+	List<Feed> findByUserId(long userId);
 	
 	List<Feed> findByFeedRefId(int feedRefId);
